voice-service/requirements.txt
# ConRad Voice Service Dependencies
faster-whisper>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
torch>=2.0.0



